+++ ios/Video/AudioSessionManager.swift
@@ -7,8 +7,17 @@ class AudioSessionManager {
     private var videoViews = NSHashTable<RCTVideo>.weakObjects()
     private var isAudioSessionActive = false
     private var remoteControlEventsActive = false
+    private var isAudioSessionManagementForcedDisabled = false
 
     private var isAudioSessionManagementDisabled: Bool {
+        if isAudioSessionManagementForcedDisabled {
+            return true
+        }
+        // If no views are registered, disable audio session management
+        if videoViews.allObjects.isEmpty {
+            return true
+        }
+
         return videoViews.allObjects.contains { view in
             return view._disableAudioSessionManagement == true
         }
@@ -37,6 +46,10 @@ class AudioSessionManager {
     }
 
     // MARK: - Public API
+
+    func setIsAudioSessionManagementForcedDisabled(disabled: Bool) {
+        isAudioSessionManagementForcedDisabled = disabled
+    }
 
     func registerView(view: RCTVideo) {
         if videoViews.contains(view) {
@@ -139,14 +152,24 @@ class AudioSessionManager {
             return view._playInBackground
         }
 
-        let canAllowMixing = !anyPlayerShowNotificationControls && !anyPlayerNeedsBackgroundPlayback
+        let anyPlayerPlaying = videoViews.allObjects.contains { view in
+            return !view.isMuted() && view._player != nil && view._player?.rate != 0
+        }
+
+        let anyPlayerWantsMixing = videoViews.allObjects.contains { view in
+            return view._mixWithOthers == "mix" || view._mixWithOthers == "duck"
+        }
+
+        let canAllowMixing = anyPlayerWantsMixing || (!anyPlayerShowNotificationControls && !anyPlayerNeedsBackgroundPlayback)
 
         if isAudioSessionManagementDisabled {
             // AUDIO SESSION MANAGEMENT DISABLED BY USER
             return
         }
 
-        if canAllowMixing {
+        if !anyPlayerPlaying {
+            options.insert(.mixWithOthers)
+        } else if canAllowMixing {
             let shouldEnableMixing = videoViews.allObjects.contains { view in
                 return view._mixWithOthers == "mix"
             }
@@ -195,7 +218,7 @@ class AudioSessionManager {
 
         do {
             try audioSession.setCategory(
-                category, mode: .moviePlayback, options: canAllowMixing ? options : []
+                category, mode: .moviePlayback, options: options
             )
 
             // Configure audio port